+++ homelab_git_management/run.sh
@@ -33,7 +33,7 @@
 
 set -e
 
-APP_VERSION="0.1.1"
+APP_VERSION="0.1.2"
 
 export APP_VERSION